Optimize simple return subs with () prototype
This has no effect on parsing, prototypes cannot effect how method calls
are parsed. That is not the point.

Adding the () here tells perl that the subs themselves take no
arguments, which optimizes away a bunch of @_ setup that otherwise takes
place. This optimization DOES effect the method calls.

This is a micro-optimization, but when we are talking about sending 10k+
events it adds up.